fix: move code to ASTHelper
