+++ src/Schema/Directives/BaseDirective.php
@@ -301,64 +301,4 @@ abstract class BaseDirective implements Directive
             $this->directiveArgs[$node->name->value] = AST::valueFromASTUntyped($node->value);
         }
     }
-
-    /**
-     * Get an instance of a directive.
-     *
-     * @api
-     */
-    public static function getDirectiveInstance(string $directiveSource, ScalarTypeDefinitionNode|ScalarTypeExtensionNode|ObjectTypeDefinitionNode|ObjectTypeExtensionNode|InterfaceTypeDefinitionNode|InterfaceTypeExtensionNode|UnionTypeDefinitionNode|UnionTypeExtensionNode|EnumTypeDefinitionNode|EnumTypeExtensionNode|InputObjectTypeDefinitionNode|InputObjectTypeExtensionNode|FieldDefinitionNode|InputValueDefinitionNode|EnumValueDefinitionNode $node): Directive
-    {
-        $directiveNode = Parser::directive($directiveSource);
-        $node->directives[] = $directiveNode;
-
-        $directiveLocator = Container::getInstance()->make(DirectiveLocator::class);
-        $directiveInstance = $directiveLocator->create($directiveNode->name->value);
-        if ($directiveInstance instanceof BaseDirective) {
-            $directiveInstance->hydrate($directiveNode, $node);
-        }
-
-        return $directiveInstance;
-    }
-
-    /**
-     * Add a directive to a field definition dynamically.
-     *
-     * @api
-     */
-    public static function addDirectiveToFieldDefinition(
-        string $directiveSource,
-        DocumentAST &$documentAST,
-        FieldDefinitionNode &$fieldDefinition,
-        ObjectTypeDefinitionNode|InterfaceTypeDefinitionNode &$parentType,
-    ): void {
-        $directiveInstance = self::getDirectiveInstance($directiveSource, $fieldDefinition);
-
-        if (!($directiveInstance instanceof FieldManipulator)) {
-            throw new DefinitionException("The dynamically added directive [{$directiveSource}] must implement the FieldManipulator interface.");
-        }
-
-        $directiveInstance->manipulateFieldDefinition($documentAST, $fieldDefinition, $parentType);
-    }
-
-    /**
-     * Add a directive to an arg definition dynamically.
-     *
-     * @api
-     */
-    public static function addDirectiveToArgDefinition(
-        string $directiveSource,
-        DocumentAST &$documentAST,
-        InputValueDefinitionNode &$argDefinition,
-        FieldDefinitionNode &$parentField,
-        ObjectTypeDefinitionNode|InterfaceTypeDefinitionNode &$parentType,
-    ): void {
-        $directiveInstance = self::getDirectiveInstance($directiveSource, $argDefinition);
-
-        if (!($directiveInstance instanceof ArgManipulator)) {
-            throw new DefinitionException("The dynamically added directive [{$directiveSource}] must implement the ArgManipulator interface.");
-        }
-
-        $directiveInstance->manipulateArgDefinition($documentAST, $argDefinition, $parentField, $parentType);
-    }
 }
