+++ tests/Unit/Schema/AST/ASTBuilderTest.php
@@ -3,26 +3,15 @@
 namespace Tests\Unit\Schema\AST;
 
 use GraphQL\Language\AST\EnumTypeDefinitionNode;
-use GraphQL\Language\AST\FieldDefinitionNode;
 use GraphQL\Language\AST\InputObjectTypeDefinitionNode;
-use GraphQL\Language\AST\InputValueDefinitionNode;
 use GraphQL\Language\AST\InterfaceTypeDefinitionNode;
-use GraphQL\Language\AST\NamedTypeNode;
 use GraphQL\Language\AST\NodeKind;
 use GraphQL\Language\AST\ObjectTypeDefinitionNode;
-use GraphQL\Language\AST\TypeDefinitionNode;
-use GraphQL\Language\Parser;
 use Illuminate\Support\Collection;
 use Nuwave\Lighthouse\Exceptions\DefinitionException;
 use Nuwave\Lighthouse\Schema\AST\ASTBuilder;
 use Nuwave\Lighthouse\Schema\AST\ASTHelper;
-use Nuwave\Lighthouse\Schema\AST\DocumentAST;
-use Nuwave\Lighthouse\Schema\Directives\BaseDirective;
 use Nuwave\Lighthouse\Schema\RootType;
-use Nuwave\Lighthouse\Support\Contracts\ArgManipulator;
-use Nuwave\Lighthouse\Support\Contracts\FieldManipulator;
-use Nuwave\Lighthouse\Support\Contracts\TypeManipulator;
-use Nuwave\Lighthouse\Support\Utils;
 use Tests\TestCase;
 
 final class ASTBuilderTest extends TestCase
@@ -291,122 +280,5 @@ final class ASTBuilderTest extends TestCase
         $this->assertCount(2, $interfaces);
         $this->assertTrue($interfaces->contains('name.value', 'Emailable'));
         $this->assertTrue($interfaces->contains('name.value', 'Nameable'));
-    }
-
-    public function testDynamicallyAddedFieldManipulatorDirective(): void
-    {
-        $directive = new class() extends BaseDirective implements FieldManipulator {
-            public static function definition(): string
-            {
-                return /** @lang GraphQL */ 'directive @foo on FIELD_DEFINITION';
-            }
-
-            public function manipulateFieldDefinition(
-                DocumentAST &$documentAST,
-                FieldDefinitionNode &$fieldDefinition,
-                ObjectTypeDefinitionNode|InterfaceTypeDefinitionNode &$parentType,
-            ):void {
-                $fieldDefinition->type = Parser::namedType('Int');
-            }
-        };
-    
-        Utils::accessProtected($this->astBuilder, 'directiveLocator')->setResolved('foo', $directive::class);
-
-        $dynamicDirective = new class() extends BaseDirective implements FieldManipulator {
-            public static function definition(): string
-            {
-                return /** @lang GraphQL */ 'directive @dynamic on FIELD_DEFINITION';
-            }
-
-            public function manipulateFieldDefinition(
-                DocumentAST &$documentAST,
-                FieldDefinitionNode &$fieldDefinition,
-                ObjectTypeDefinitionNode|InterfaceTypeDefinitionNode &$parentType,
-            ):void {
-                $this->addDirectiveToFieldDefinition('@foo', $documentAST, $fieldDefinition, $parentType);
-            }
-        };
-
-        Utils::accessProtected($this->astBuilder, 'directiveLocator')->setResolved('dynamic', $dynamicDirective::class);
-
-        $this->schema = /** @lang GraphQL */ '
-        type Query {
-            foo: String @dynamic
-        }
-        ';
-        $documentAST = $this->astBuilder->documentAST();
-
-        $queryType = $documentAST->types[RootType::QUERY];
-        assert($queryType instanceof ObjectTypeDefinitionNode);
-
-        $fieldType = $queryType->fields[0];
-        assert($fieldType instanceof FieldDefinitionNode);
-
-        $typeType = $fieldType->type;
-        assert($typeType instanceof NamedTypeNode);
-
-        $this->assertEquals($typeType->name->value, 'Int');
-    }
-
-    public function testDynamicallyAddedArgManipulatorDirective(): void
-    {
-        $directive = new class() extends BaseDirective implements ArgManipulator {
-            public static function definition(): string
-            {
-                return /** @lang GraphQL */ 'directive @foo on ARGUMENT_DEFINITION';
-            }
-
-            public function manipulateArgDefinition(
-                DocumentAST &$documentAST,
-                InputValueDefinitionNode &$argDefinition,
-                FieldDefinitionNode &$parentField,
-                ObjectTypeDefinitionNode|InterfaceTypeDefinitionNode &$parentType,
-            ): void
-            {
-                $argDefinition->type = Parser::namedType('Int');
-            }
-        };
-    
-        Utils::accessProtected($this->astBuilder, 'directiveLocator')->setResolved('foo', $directive::class);
-
-        $dynamicDirective = new class() extends BaseDirective implements ArgManipulator {
-            public static function definition(): string
-            {
-                return /** @lang GraphQL */ 'directive @dynamic on ARGUMENT_DEFINITION';
-            }
-
-            public function manipulateArgDefinition(
-                DocumentAST &$documentAST,
-                InputValueDefinitionNode &$argDefinition,
-                FieldDefinitionNode &$parentField,
-                ObjectTypeDefinitionNode|InterfaceTypeDefinitionNode &$parentType,
-            ): void
-            {
-                $this->addDirectiveToArgDefinition('@foo', $documentAST, $argDefinition, $parentField, $parentType);
-            }
-        };
-
-        Utils::accessProtected($this->astBuilder, 'directiveLocator')->setResolved('dynamic', $dynamicDirective::class);
-
-        $this->schema = /** @lang GraphQL */ '
-        type Query {
-            foo( name: String @dynamic ): String
-        }
-        ';
-        $documentAST = $this->astBuilder->documentAST();
-
-        $queryType = $documentAST->types[RootType::QUERY];
-        assert($queryType instanceof ObjectTypeDefinitionNode);
-
-        $fieldType = $queryType->fields[0];
-        assert($fieldType instanceof FieldDefinitionNode);
-
-        $argumentType = $fieldType->arguments[0];
-        assert($argumentType instanceof InputValueDefinitionNode);
-
-        $typeType = $argumentType->type;
-        assert($typeType instanceof NamedTypeNode);
-
-        $this->assertEquals($typeType->name->value, 'Int');
     }
 }
